+++ src/controls/controlRendering.ts
@@ -11,13 +11,15 @@ export type ControlRenderingStyleOverride = Partial<
     | 'cornerStrokeColor'
     | 'cornerDashArray'
     | 'transparentCorners'
+    | 'borderColor'
+    | 'borderDashArray'
   >
 >;
 
 export type ControlRenderer = (
   ctx: CanvasRenderingContext2D,
-  left: number,
-  top: number,
+  x: number,
+  y: number,
   styleOverride: ControlRenderingStyleOverride,
   fabricObject: FabricObject
 ) => void;
@@ -28,16 +30,16 @@ export type ControlRenderer = (
  * cornerColor, cornerStrokeColor
  * plus the addition of offsetY and offsetX.
  * @param {CanvasRenderingContext2D} ctx context to render on
- * @param {Number} left x coordinate where the control center should be
- * @param {Number} top y coordinate where the control center should be
+ * @param {number} x control center x
+ * @param {number} y control center y
  * @param {Object} styleOverride override for FabricObject controls style
  * @param {FabricObject} fabricObject the fabric object for which we are rendering controls
  */
 export function renderCircleControl(
   this: Control,
   ctx: CanvasRenderingContext2D,
-  left: number,
-  top: number,
+  x: number,
+  y: number,
   styleOverride: ControlRenderingStyleOverride,
   fabricObject: FabricObject
 ) {
@@ -53,9 +55,7 @@ export function renderCircleControl(
     stroke =
       !transparentCorners &&
       (styleOverride.cornerStrokeColor || fabricObject.cornerStrokeColor);
-  let myLeft = left,
-    myTop = top,
-    size;
+  let size: number;
   ctx.save();
   ctx.fillStyle = styleOverride.cornerColor || fabricObject.cornerColor || '';
   ctx.strokeStyle =
@@ -64,18 +64,18 @@ export function renderCircleControl(
   if (xSize > ySize) {
     size = xSize;
     ctx.scale(1.0, ySize / xSize);
-    myTop = (top * xSize) / ySize;
+    y *= xSize / ySize;
   } else if (ySize > xSize) {
     size = ySize;
     ctx.scale(xSize / ySize, 1.0);
-    myLeft = (left * ySize) / xSize;
+    x *= ySize / xSize;
   } else {
     size = xSize;
   }
   // this is still wrong
   ctx.lineWidth = 1;
   ctx.beginPath();
-  ctx.arc(myLeft, myTop, size / 2, 0, twoMathPi, false);
+  ctx.arc(x, y, size / 2, 0, twoMathPi, false);
   ctx[methodName]();
   if (stroke) {
     ctx.stroke();
@@ -89,16 +89,16 @@ export function renderCircleControl(
  * cornerColor, cornerStrokeColor
  * plus the addition of offsetY and offsetX.
  * @param {CanvasRenderingContext2D} ctx context to render on
- * @param {Number} left x coordinate where the control center should be
- * @param {Number} top y coordinate where the control center should be
+ * @param {number} x control center x
+ * @param {number} y control center y
  * @param {Object} styleOverride override for FabricObject controls style
  * @param {FabricObject} fabricObject the fabric object for which we are rendering controls
  */
 export function renderSquareControl(
   this: Control,
   ctx: CanvasRenderingContext2D,
-  left: number,
-  top: number,
+  x: number,
+  y: number,
   styleOverride: ControlRenderingStyleOverride,
   fabricObject: FabricObject
 ) {